added configure option for building private headers
there have been a number of reported issues with the decision to
default to using QObjectPrivate in the library. in cases where users
are using the distributed SDK (mostly OSX and Windows), the private
headers are not installed and therefore it's impossible to build
qjsonrpc in that configuration. this change reverts to using no
private headers by default, but giving the option "private-headers"
to opt-in to that build flavor